+++ src/components/myPage/ShowUserInfo/MyCart.tsx
@@ -201,9 +201,10 @@ export const MyCart: FC<MyCartProps> = ({
     // 스팟 삭제
     const deleteSpot = async (mno: number, pno: number, fno: number) => {
         try {
-            await deleteCart(mno, pno, fno)
-            fetchData()
-            alert('스팟 삭제!')
+            if (window.confirm('삭제 하시겠습니까?')) {
+                await deleteCart(mno, pno, fno)
+                fetchData()
+            }
         } catch (error) {
             console.log(error)
         }
@@ -232,7 +233,7 @@ export const MyCart: FC<MyCartProps> = ({
                             .map(folderInfo => (
                                 <div
                                     key={folderInfo.fno}
-                                    className={`h-full relative cursor-pointer justify-center flex items-center p-2 w-full text-black text-sm duration-150 border border-b-0 rounded-t-2xl ${
+                                    className={`max-w-[250px] h-full relative cursor-pointer justify-center flex items-center p-2 w-full text-black text-sm duration-150 border border-b-0 rounded-t-2xl ${
                                         selectedFno === folderInfo.fno
                                             ? 'bg-lightGreen text-white bg-opacity-80'
                                             : 'bg-white'
